Rename common transducers package 'transducers'
