+++ base.js
@@ -1,4 +1,4 @@
 'use strict'
 var core = require('./core'),
     merge = core.util.objectMerge
-module.exports = merge(merge({}, core), require('./common'))
+module.exports = merge(merge({}, core), require('./transducers'))
