+++ transducers/drop.js
