+++ transducers/index.js
