+++ transducers/map.js
