+++ transducers/mapcat.js
