+++ transducers/partitionAll.js
