+++ transducers/partitionBy.js
