+++ transducers/remove.js
