+++ transducers/take.js
