+++ transducers/takeWhile.js
